// src/app/core/models/driver.model.ts
export interface Driver {
    docId?: string;
    driverName: string;
    driverCode: string;
    driverNumber?: string;
    mobileNumber?: string;
    altMobileNumber?: string;
    address?: string;
    driverLocation?: string;
    bloodGroup?: string;
    licenseNumber?: string;
    state?: string;
    district?: string;
    pinCode?: string;
    driverGrade?: string;
    dateOfJoining?: string;
    dateOfResigning?: string;
    status?: string;
    id?: string;
    firstName?: string;
    lastName?: string;
    fullName?: string;
    aadhaarNumber?: string;
    dateOfBirth?: string;
    driverType?: string;
    path?: string;
}